# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'lciii-scaffolding'","description":"'LCII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'LCII Initial Scaffolding'"}],"paths":{"/countries":{"get":{"tags":["country-controller"],"operationId":"getAllCountries","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}},{"name":"code","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}},"post":{"tags":["country-controller"],"operationId":"createCountry","requestBody":{"content":{"application/json":{"schema":{"type":"integer","format":"int64"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"/countries/{language}/language":{"get":{"tags":["country-controller"],"operationId":"getCountrysByLanguage","parameters":[{"name":"language","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"/countries/{continent}/continent":{"get":{"tags":["country-controller"],"operationId":"getCountrysByContinent","parameters":[{"name":"continent","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"/countries/most-borders":{"get":{"tags":["country-controller"],"operationId":"getCountryWithMostBorders","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"components":{"schemas":{"CountryDTO":{"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"}}}}}}